requirements.txt
pandas
numpy
matplotlib
PyYAML
paramiko
scp
seaborn
